requirements.txt
torch==2.10.0
torchvision==0.21.0
numpy==2.2.6
matplotlib==3.10.3
seaborn==0.13.2
scikit-learn==1.7.0
Pillow==11.2.1
tqdm==4.67.1
pandas==2.2.3
pyyaml==6.0.2
